Move VendorAPI docs under APIDocs namespace
As we add other APIDocs we can put them under this namespace too
