--- app/controllers/api_docs/openapi_controller.rb
@@ -1,7 +1,0 @@
-module APIDocs
-  class OpenapiController < APIDocsController
-    def spec
-      render plain: VendorAPISpecification.as_yaml, content_type: 'text/yaml'
-    end
-  end
-end
