--- app/controllers/api_docs/reference_controller.rb
@@ -1,7 +1,0 @@
-module APIDocs
-  class ReferenceController < APIDocsController
-    def reference
-      @api_reference = APIReference.new
-    end
-  end
-end
