+++ app/controllers/api_docs/vendor_api_docs/openapi_controller.rb
@@ -1,0 +1,9 @@
+module APIDocs
+  module VendorAPIDocs
+    class OpenapiController < APIDocsController
+      def spec
+        render plain: VendorAPISpecification.as_yaml, content_type: 'text/yaml'
+      end
+    end
+  end
+end
