+++ app/controllers/api_docs/vendor_api_docs/reference_controller.rb
@@ -1,0 +1,9 @@
+module APIDocs
+  module VendorAPIDocs
+    class ReferenceController < APIDocsController
+      def reference
+        @api_reference = APIReference.new(VendorAPISpecification.as_hash)
+      end
+    end
+  end
+end
