+++ app/helpers/navigation_items.rb
@@ -86,14 +86,14 @@ class NavigationItems
       end
     end
 
-    def for_api_docs(current_controller)
+    def for_vendor_api_docs(current_controller)
       [
         NavigationItem.new('Home', api_docs_home_path, is_active_action(current_controller, 'home')),
-        NavigationItem.new(t('page_titles.api_docs.usage'), api_docs_usage_path, is_active_action(current_controller, 'usage')),
-        NavigationItem.new(t('page_titles.api_docs.reference'), api_docs_reference_path, is_active_action(current_controller, 'reference')),
-        NavigationItem.new(t('page_titles.api_docs.release_notes'), api_docs_release_notes_path, is_active_action(current_controller, 'release_notes')),
-        NavigationItem.new(t('page_titles.api_docs.lifecycle'), api_docs_lifecycle_path, is_active_action(current_controller, 'lifecycle')),
-        NavigationItem.new(t('page_titles.api_docs.help'), api_docs_help_path, is_active_action(current_controller, 'help')),
+        NavigationItem.new(t('page_titles.api_docs.vendor_api_docs.usage'), api_docs_usage_path, is_active_action(current_controller, 'usage')),
+        NavigationItem.new(t('page_titles.api_docs.vendor_api_docs.reference'), api_docs_reference_path, is_active_action(current_controller, 'reference')),
+        NavigationItem.new(t('page_titles.api_docs.vendor_api_docs.release_notes'), api_docs_release_notes_path, is_active_action(current_controller, 'release_notes')),
+        NavigationItem.new(t('page_titles.api_docs.vendor_api_docs.lifecycle'), api_docs_lifecycle_path, is_active_action(current_controller, 'lifecycle')),
+        NavigationItem.new(t('page_titles.api_docs.vendor_api_docs.help'), api_docs_help_path, is_active_action(current_controller, 'help')),
       ]
     end
