+++ app/views/api_docs/vendor_api_docs/pages/lifecycle.html.erb
