+++ app/views/api_docs/vendor_api_docs/pages/rendered_markdown_template.html.erb
@@ -1,0 +1,9 @@
+<%= content_for :title, t("page_titles.api_docs.vendor_api_docs.#{@page_name}") %>
+
+<div class="govuk-grid-row">
+  <div class="govuk-grid-column-two-thirds">
+    <h1 class="govuk-heading-xl"><%= t("page_titles.api_docs.vendor_api_docs.#{@page_name}") %></h1>
+
+    <%= @converted_markdown %>
+  </div>
+</div>
