+++ app/views/api_docs/vendor_api_docs/pages/when_emails_are_sent.html.erb
