+++ app/views/api_docs/vendor_api_docs/reference/reference.html.erb
@@ -1,4 +1,4 @@
-<h1 class="govuk-heading-xl"><%= t('page_titles.api_docs.reference') %></h1>
+<h1 class="govuk-heading-xl"><%= t('page_titles.api_docs.vendor_api_docs.reference') %></h1>
 
 <h2 class="app-contents-list__title">Contents</h2>
 
@@ -112,4 +112,3 @@
 <%= render(APIDocs::APIReferenceComponent.new(@api_reference)) %>
 
 <%= render(APIDocs::FieldLengthsComponent.new(@api_reference)) %>
-
