+++ config/locales/en.yml
@@ -175,13 +175,14 @@ en:
       withdraw: Are you sure you want to withdraw this course choice?
       withdrawal_feedback: Course choice withdrawn
     api_docs:
-      home: Apply for teacher training
-      usage: Usage scenarios
-      reference: API reference
-      help: Get help
-      release_notes: Release notes
-      lifecycle: Lifecycle
-      alpha_release_notes: Alpha release notes
+      vendor_api_docs:
+        home: Apply for teacher training
+        usage: Usage scenarios
+        reference: API reference
+        help: Get help
+        release_notes: Release notes
+        lifecycle: Lifecycle
+        alpha_release_notes: Alpha release notes
     provider:
       confirm_offer: Check and confirm offer
       confirm_reject: Check feedback and reject application
